+++ CHANGES.rst
@@ -4,8 +4,9 @@ Changelog
 1.1 (unreleased)
 ----------------
 
-- Nothing changed yet.
-
+- Also support IEventAccessor in get_location and fix location not beeing
+  displayed on ``@@event_listing`` views.
+  [thet]
 
 1.0 (2014-04-30)
 ----------------